Raise error when Gemfile.lock not present
This change fixes a bug: Previously, when bundler audit
emitted a `warning: no gemfile.lock found` json, it would
cause the snapshot to get stuck in started state. All steps
would finish, but Builder would not send a `done` message.

Context: Bundler-audit reports some repo-wide issues
that do not belong to any single file. In this case for
instance, the issue is that no Gemfile.lock exists.

Since Code Climate does not presently provide an interface
for handling repo-wide issues, the engine should blow up
when this warning is present.

In the future, we want to more smoothly handle reports of
orphan issues.