Emit a unique fingerprint for issues
Previously this engine did not emit a fingerprint for any issues which
resulted in Code Climate assigning a default fingerprint. As the
the default fingerprint is just a concatenation of the path of the file
and the check name, only two fingerprints were possible depending on
which check emitted the issue.

Instead, use details from the issue to calculate a fingerprint. For an
insecure source, use the source and for a gem use the gem name and the
advisory ID.