Respect include paths
Despite being effectively a single-file engine, `bundler-audit` should
still respect include paths, and not run analysis if `Gemfile.lock` is
not in the include paths. The primary use case affected is analyzing
specific files with the CLI (e.g. `codeclimate analyze
app/models/foo.rb`).

This change keeps the existing "raise error if file does not exist"
check (which seems reasonable to me in all cases), but checks for
whether it would be considered as part of the include paths before
running analysis.